--- lib/http/model_controller.dart
@@ -1,70 +1,0 @@
-part of aqueduct;
-
-/// A partial class for implementing an [HTTPController] that has a few conveniences
-/// for handling [ManagedObject] objects.
-///
-/// Intended to be subclassed. Instances of [QueryController] are [HTTPController]s whose
-/// type argument indicates the type of [ManagedObject] object they are responsible for executing [Query]s
-/// on. [QueryController]s expose a [query] property that is pre-populated with data from the incoming [Request].
-///
-/// Prior to executing a controller method in subclasses of [QueryController], the [query]
-/// will have the following attributes under the following conditions:
-///
-/// 1. The Query will always have a type argument that matches [T].
-/// 2. If the request contains a path variable, the first path variable (reading from left to right) will be used to as a matcher
-/// on the entity's primary key. For example, a [Request] with the path /users/:id and a type argument of User (whose primary key is userID) will generate a Query as though:
-///
-///         query = new ModelQuery<User>()..userID = whereEqualTo(req.path["id"]);
-///
-/// 3. If the [Request] contains a body, it will be decoded per the [acceptedContentTypes] and deserialized into the [query]'s [values] property via [readMap].
-abstract class QueryController<T extends ManagedObject> extends HTTPController {
-
-  /// Create an instance of [QueryController]. By default, [context] is the [defaultContext].
-  QueryController([ManagedContext context]) : super() {
-    query = new Query<T>(context: context ?? ManagedContext.defaultContext);
-  }
-
-  /// A query representing the values received from the [request] being processed.
-  ///
-  /// You may execute this [query] as is or modify it.
-  Query<T> query;
-
-  @override
-  Future<RequestControllerEvent> willProcessRequest(Request req) async {
-    if (req.path.orderedVariableNames.length > 0) {
-      var firstVarName = req.path.orderedVariableNames.first;
-      var idValue = req.path.variables[firstVarName];
-
-      if (idValue != null) {
-        var primaryKeyDesc = query.entity.attributes[query.entity.primaryKey];
-        if (primaryKeyDesc.isAssignableWith(idValue)) {
-          query.matchOn[query.entity.primaryKey] = idValue;
-        } else if (primaryKeyDesc.type == ManagedPropertyType.bigInteger || primaryKeyDesc.type == ManagedPropertyType.integer) {
-          try {
-            query.matchOn[query.entity.primaryKey] = int.parse(idValue);
-          } on FormatException {
-            var errorMessage = "Expected integer value for ModelController on ${query.entity}, but $idValue was not able to be parsed to an integer.";
-
-            return new Response.notFound(body: {"error" : errorMessage});
-          }
-        } else {
-          var errorMessage = "ID Value $idValue is not assignable for ModelController on ${query.entity}, expected value of type ${primaryKeyDesc.type}";
-
-          return new Response.notFound(body: {"error" : errorMessage});
-        }
-      }
-    }
-
-    return super.willProcessRequest(req);
-  }
-
-  @override
-  void didDecodeRequestBody(dynamic body) {
-    var bodyMap = body as Map<String, dynamic>;
-    var reflectedModel = reflectClass(T).newInstance(new Symbol(""), []);
-    query.values = reflectedModel.reflectee as T;
-    query.values.readMap(bodyMap);
-
-    query.values.removePropertyFromBackingMap(query.values.entity.primaryKey);
-  }
-}
