+++ test/auth/auth_code_controller_test.dart
@@ -243,7 +243,7 @@ class TestSink extends RequestSink {
 
   AuthServer authServer;
 
-  void addRoutes() {
+  void setupRouter(Router router) {
     router
         .route("/auth/code")
         .generate(() => new AuthCodeController(authServer));
