+++ test/base/application_test.dart
@@ -111,7 +111,7 @@ class CrashingTestSink extends RequestSink {
     }
   }
 
-  void addRoutes() {
+  void setupRouter(Router router) {
     if (options["crashIn"] == "addRoutes") {
       throw new TestException("addRoutes");
     }
@@ -129,7 +129,7 @@ class CrashingTestSink extends RequestSink {
 class TestSink extends RequestSink {
   TestSink(Map<String, dynamic> opts) : super(opts);
 
-  void addRoutes() {
+  void setupRouter(Router router) {
     router.route("/t").generate(() => new TController());
     router.route("/r").generate(() => new RController());
