+++ test/base/document_test.dart
@@ -428,7 +428,7 @@ class TestSink extends RequestSink implements AuthServerDelegate<TestUser, Token
 
   AuthServer<TestUser, Token, AuthCode> authServer;
 
-  void addRoutes() {
+  void setupRouter(Router router) {
     router.route("/auth/code").pipe(new Authorizer(authServer, strategy: AuthStrategy.client)).generate(() => new AuthCodeController(authServer));
     router.route("/auth/token").pipe(new Authorizer(authServer, strategy: AuthStrategy.client)).generate(() => new AuthController(authServer));
     router.route("/t[/:id[/:notID]]").pipe(new Authorizer(authServer)).generate(() => new TController());
