+++ test/test_project/lib/src/wildfire_sink.dart
@@ -24,7 +24,7 @@ class WildfireSink extends RequestSink {
   WildfireConfiguration configuration;
 
   @override
-  void addRoutes() {
+  void setupRouter(Router router) {
     router
         .route("/auth/token")
         .pipe(new Authorizer(authenticationServer, strategy: AuthStrategy.client))
